test: fold the day-index suites into events.test.ts
`day-index.test.ts` named a module that does not exist. Everything it
imported came from `events.ts`, which therefore had two test files with two
separate builder sets and nothing indicating which held what — against
AGENTS.md's "never create new test files".

The five describes move into `events.test.ts` unchanged, the duplicated
module-level `afterEach` is dropped, and a file docstring states the file's
scope. `@ilamy/utils` still runs 123 tests, now across 2 files instead of 3.

Requested in review of #241.
